requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
PyPDF2>=3.0.0
openai>=1.0.0
python-dotenv>=0.19.0
pydantic>=2.0.0
